rename sub directory names
