+++ packages/base/src/formbuilder/index.ts
@@ -1,6 +1,6 @@
 export * from './objectform/baseform';
-export * from './objectform/sourceforms';
-export * from './objectform/layerforms';
-export * from './objectform/sourceforms';
-export * from './objectform/layerforms';
+export * from './objectform/source';
+export * from './objectform/layer';
+export * from './objectform/source';
+export * from './objectform/layer';
 export * from './creationform';
